Exercises 4.11-4.12

+++ part4/bloglist/controllers/blogs.js
@@ -1,58 +1,67 @@
-const blogRouter = require('express').Router()
-const Blog = require('../models/blog')
-const User = require('../models/user')
-const jwt = require('jsonwebtoken')
+const blogRouter = require('express').Router();
+const Blog = require('../models/blog');
+const User = require('../models/user');
+const jwt = require('jsonwebtoken');
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
-  response.json(blogs)
-})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 });
+  response.json(blogs);
+});
 
 blogRouter.post('/', async (request, response) => {
-  const body = request.body
-  
+  const body = request.body;
+
   if (!request.token) {
-    return response.status(401).json({ error: 'token missing' })
+    return response.status(401).json({ error: 'token missing' });
   }
-  
-  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
   if (!decodedToken.id) {
-    return response.status(401).json({ error: 'token invalid' })
+    return response.status(401).json({ error: 'token invalid' });
   }
-  
-  const user = await User.findById(decodedToken.id)
-  
+
+  const user = await User.findById(decodedToken.id);
+
   const blog = new Blog({
     title: body.title,
     author: body.author,
     url: body.url,
     likes: body.likes,
-    user: user._id
-  })
-  
-  const savedBlog = await blog.save()
-  
-  user.blogs = user.blogs.concat(savedBlog._id)
-  await user.save()
-  
-  response.status(201).json(savedBlog)
-})
+    user: user._id,
+  });
+
+  const savedBlog = await blog.save();
+
+  user.blogs = user.blogs.concat(savedBlog._id);
+  await user.save();
+
+  response.status(201).json(savedBlog);
+});
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
-})
+  if (!request.token) {
+    return response.status(401).json({ error: 'token missing' });
+  }
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
+
+  await Blog.findByIdAndDelete(request.params.id);
+  response.status(204).end();
+});
 
 blogRouter.put('/:id', async (request, response) => {
-  const { title, author, url, likes } = request.body
-  
+  const { title, author, url, likes } = request.body;
+
   const updatedBlog = await Blog.findByIdAndUpdate(
     request.params.id,
     { title, author, url, likes },
     { new: true, runValidators: true, context: 'query' }
-  )
-  
-  response.json(updatedBlog)
-})
+  );
 
-module.exports = blogRouter
+  response.json(updatedBlog);
+});
+
+module.exports = blogRouter;
